Flask-App/requirements.txt
tensorflow
keras
pygeohash
pandas
matplotlib
sklearn
pillow
flask
django
